requirements.txt
bullet
python-box
pydantic